Fixes wrong subtraction of prompt weigthing field
In the case of the delayed weighting field (defined according to the
weighting vector by W. Riegler in DOI 10.1016/j.nima.2019.06.056), the
delayed component is calculated separatly from the prompt component (see
equations (17) and (18) in the mentionned paper).

Thus the prompt component does not need to be subtracted from the
delayed component.